PI-9129 : add support of Microsoft service to get subscription provisioning state on PartnerCenterAdmin.
