+++ src/main/java/org/springframework/social/partnercenter/api/order/AdminOrderOperations.java
@@ -1,0 +1,8 @@
+package org.springframework.social.partnercenter.api.order;
+
+import org.springframework.http.ResponseEntity;
+import org.springframework.social.partnercenter.api.order.subscription.SubscriptionProvisioningState;
+
+public interface AdminOrderOperations extends OrderOperations {
+	ResponseEntity<SubscriptionProvisioningState> getSubscriptionProvisioningState(String customerTenantId, String subscriptionId);
+}
