+++ src/main/java/org/springframework/social/partnercenter/api/order/impl/OrderTemplate.java
@@ -8,7 +8,6 @@ import org.springframework.social.partnercenter.api.PartnerCenterResponse;
 import org.springframework.social.partnercenter.api.order.Order;
 import org.springframework.social.partnercenter.api.order.OrderOperations;
 import org.springframework.social.partnercenter.api.order.request.CreateOrderRequest;
-import org.springframework.social.partnercenter.api.order.subscription.SubscriptionProvisioningState;
 import org.springframework.social.partnercenter.http.client.RestResource;
 
 public class OrderTemplate extends PagingResourceTemplate<Order> implements OrderOperations {
@@ -50,12 +49,5 @@ public class OrderTemplate extends PagingResourceTemplate<Order> implements Orde
 	@Override
 	protected String getProviderId() {
 		return PartnerCenter.PROVIDER_ID;
-	}
-
-	@Override
-	public ResponseEntity<SubscriptionProvisioningState> getSubscriptionProvisioningState(String customerTenantId, String subscriptionId) {
-		return restResource.request()
-				.pathSegment(customerTenantId, "subscriptions", subscriptionId, "provisioningstatus")
-				.get(SubscriptionProvisioningState.class);
 	}
 }
